Lets try that again.  So, select count(*) returns a long and Projections.rowCount returns an integer.

+++ com.zutubi.pulse.master/src/java/com/zutubi/pulse/master/model/persistence/hibernate/HibernateSearchQuery.java
@@ -192,7 +192,7 @@ public class HibernateSearchQuery<T> implements Serializable
      */
     public long count()
     {
-        return (Long)getHibernateTemplate().execute(new HibernateCallback()
+        return (Integer)getHibernateTemplate().execute(new HibernateCallback()
         {
             public Object doInHibernate(Session session) throws HibernateException
             {
